+++ frontend/app/src/components/staking/liquity/LiquityStakingDetails.vue
@@ -26,9 +26,13 @@
       </v-col>
     </v-row>
 
-    <v-row v-if="stakingList.length > 0" class="mt-8">
-      <v-col v-for="stake in stakingList" :key="stake.asset">
-        <liquity-stake :stake="stake" />
+    <v-row>
+      <v-col md="6" cols="12">
+        <liquity-pools :pool="stakingPoolsList" />
+      </v-col>
+
+      <v-col md="6" cols="12">
+        <liquity-stake :stakings="stakingList" />
       </v-col>
     </v-row>
 
@@ -45,22 +49,28 @@ import { AssetBalance } from '@rotki/common';
 import { GeneralAccount } from '@rotki/common/lib/account';
 import { Blockchain } from '@rotki/common/lib/blockchain';
 import {
+  LiquityPoolDetail,
+  LiquityPoolDetails,
   LiquityStaking,
   LiquityStakingEvent,
   LiquityStakingEvents
 } from '@rotki/common/lib/liquity';
+import { ComputedRef } from 'vue';
 import RefreshButton from '@/components/helper/RefreshButton.vue';
+import LiquityPools from '@/components/staking/liquity/LiquityPools.vue';
 import LiquityStake from '@/components/staking/liquity/LiquityStake.vue';
 import { isSectionLoading } from '@/composables/common';
 import { LiquityStakeEvents } from '@/premium/premium';
 import { useLiquityStore } from '@/store/defi/liquity';
 import { Section } from '@/types/status';
+import { balanceSum } from '@/utils/calculation';
+import { uniqueStrings } from '@/utils/data';
 
 const selectedAccount = ref<GeneralAccount | null>(null);
 
 const liquityStore = useLiquityStore();
-const { staking, stakingEvents } = toRefs(liquityStore);
-const { fetchStaking, fetchStakingEvents } = liquityStore;
+const { staking, stakingEvents, stakingPools } = toRefs(liquityStore);
+const { fetchStaking, fetchStakingEvents, fetchPools } = liquityStore;
 
 const loading = isSectionLoading(Section.DEFI_LIQUITY_STAKING);
 const eventsLoading = isSectionLoading(Section.DEFI_LIQUITY_STAKING_EVENTS);
@@ -68,9 +78,10 @@ const chains = [Blockchain.ETH];
 
 const { tc } = useI18n();
 
-const stakingList = computed(() => {
+const stakingList: ComputedRef<AssetBalance[]> = computed(() => {
   const staked: Record<string, AssetBalance> = {};
   const stake = get(staking) as LiquityStaking;
+
   for (const address in stake) {
     const account = get(selectedAccount);
     if (account && account.address !== address) {
@@ -93,7 +104,7 @@ const stakingList = computed(() => {
   return Object.values(staked);
 });
 
-const stakingEventsList = computed(() => {
+const stakingEventsList: ComputedRef<LiquityStakingEvent[]> = computed(() => {
   const allEvents = get(stakingEvents) as LiquityStakingEvents;
   const events: LiquityStakingEvent[] = [];
   for (const address in allEvents) {
@@ -106,12 +117,47 @@ const stakingEventsList = computed(() => {
   return events;
 });
 
+const stakingPoolsList: ComputedRef<LiquityPoolDetail | null> = computed(() => {
+  const allPools = get(stakingPools) as LiquityPoolDetails;
+  let pools: LiquityPoolDetail | null = null;
+
+  for (const address in allPools) {
+    const account = get(selectedAccount);
+    if (account && account.address !== address) {
+      continue;
+    }
+    const pool = allPools[address];
+    if (pools === null) {
+      pools = { ...pool };
+    } else {
+      pools.gains = {
+        asset: pools.gains.asset,
+        ...balanceSum(pools.gains, pool.gains)
+      };
+      pools.rewards = {
+        asset: pools.rewards.asset,
+        ...balanceSum(pools.rewards, pool.rewards)
+      };
+      pools.deposited = {
+        asset: pools.deposited.asset,
+        ...balanceSum(pools.deposited, pool.deposited)
+      };
+    }
+  }
+  return pools;
+});
+
 const availableAddresses = computed(() => {
-  return [...Object.keys(get(staking)), ...Object.keys(get(stakingEvents))];
+  return [
+    ...Object.keys(get(staking)),
+    ...Object.keys(get(stakingEvents)),
+    ...Object.keys(get(stakingPools))
+  ].filter(uniqueStrings);
 });
 
 const refresh = async () => {
   await fetchStaking(true);
   await fetchStakingEvents(true);
+  await fetchPools(true);
 };
 </script>