+++ docs/changelog.rst
@@ -3,6 +3,7 @@ Changelog
 =========
 
 * :release:`1.26.1 <2022-11-04>`
+* :feature:`5102` Show stability pool to liquity module.
 * :feature:`5080` For custom assets with custom price there should no longer be any double conversion. So 1 euro should always be one euro.
 * :feature:`5046` Users who deleted important assets from their rotki instance will now have a fallback and won't get their rotki stuck.
 * :bug:`5041` Add an option to only show ignored assets in assets table.